Force Destination setup tests on creation
